[Mock테스트 3단계] 계산 기록 피드백 - 의미없는 인터페이스 삭제
domain/src/main/java/edu/nextstep/camp/calculator/domain/CalculatorMemory.kt
Outdated
interface CalculatorMemory {
    fun loadExpressionRecords(): List<String>
    fun saveExpressionRecord(expression: Expression, result: Int)
}

@laco-dev laco-dev 12 hours ago
CalculatorMemory를 인터페이스로 만들어주신 이유가 있을까요?
서로 다른 구현체를 넣는 로직이나 인터페이스를 이용한 다른 전략적인 요소가 보이지 않습니다.
오히려 Impl.XXX 형태로 직접적으로 접근하는 코드가 많아져서 의미가 없게 되었습니다.

CalculatorMemory는 실제로 인스턴스를 만들면서 테스트해야 하는게 목적에 맞다고 생각이 드네요.
