+++ app/src/test/java/edu/nextstep/camp/calculator/MainPresenterTest.kt
@@ -1,6 +1,6 @@
 package edu.nextstep.camp.calculator
 
-import edu.nextstep.camp.calculator.domain.CalculatorMemoryImpl
+import edu.nextstep.camp.calculator.domain.CalculatorMemory
 import edu.nextstep.camp.calculator.domain.Expression
 import edu.nextstep.camp.calculator.domain.Operator
 import io.mockk.mockk
@@ -121,10 +121,10 @@ internal class MainPresenterTest {
     @Test
     internal fun `계산 기록이 표시되지 않은 상태에서 계산 기록을 활성화하면 계산 기록이 노출된다`() {
         //given
-        val expected: List<String> = listOf(CalculatorMemoryImpl.SAVE_FORMAT.format("1 + 32", "33"))
+        val expected: List<String> = listOf(CalculatorMemory.SAVE_FORMAT.format("1 + 32", "33"))
         presenter = MainPresenter(
             view = view,
-            calculatorMemory = CalculatorMemoryImpl(arrayListOf(CalculatorMemoryImpl.SAVE_FORMAT.format("1 + 32", "33"))),
+            calculatorMemory = CalculatorMemory(arrayListOf(CalculatorMemory.SAVE_FORMAT.format("1 + 32", "33"))),
             isRecordsMode = false)
 
         //when
@@ -141,7 +141,7 @@ internal class MainPresenterTest {
         presenter = MainPresenter(
             view = view,
             expression = Expression(listOf(1, Operator.Plus)),
-            calculatorMemory = CalculatorMemoryImpl(arrayListOf(CalculatorMemoryImpl.SAVE_FORMAT.format("1 + 32", "33"))),
+            calculatorMemory = CalculatorMemory(arrayListOf(CalculatorMemory.SAVE_FORMAT.format("1 + 32", "33"))),
             isRecordsMode = true)
 
         //when
@@ -154,7 +154,7 @@ internal class MainPresenterTest {
     @Test
     internal fun `연산을 할 때 마다 계산 기록이 저장되어야 한다`() {
         //given
-        val memory: CalculatorMemoryImpl = mockk(relaxed = true)
+        val memory: CalculatorMemory = mockk(relaxed = true)
         presenter = MainPresenter(
             view = view,
             expression = Expression(listOf(1, Operator.Plus, 32)),
