+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/CalculatorMemory.kt
@@ -1,6 +1,21 @@
 package edu.nextstep.camp.calculator.domain
 
-interface CalculatorMemory {
-    fun loadExpressionRecords(): List<String>
-    fun saveExpressionRecord(expression: Expression, result: Int)
+class CalculatorMemory(
+    private val calcHistory: ArrayList<String> = arrayListOf()
+) {
+    companion object {
+        const val SAVE_FORMAT = "%s\n= %s\n"
+    }
+
+    fun saveExpressionRecord(expression: Expression, result: Int) {
+        calcHistory.add(SAVE_FORMAT.format(expression, result))
+    }
+
+    fun loadExpressionRecords(): List<String> {
+        return calcHistory.toList()
+    }
+
+    override fun toString(): String {
+        return calcHistory.joinToString("\n") { it }
+    }
 }
