+++ domain/src/test/java/edu/nextstep/camp/calculator/domain/CalculatorMemoryTest.kt
@@ -9,7 +9,7 @@ internal class CalculatorMemoryTest {
     @Test
     internal fun `완성된 수식이 주어질 때 계산 기록이 저장되어야 한다`() {
         //given
-        calculatorMemory = CalculatorMemoryImpl()
+        calculatorMemory = CalculatorMemory()
         val expression = Expression(listOf(3, Operator.Plus, 5))
         val result = 8
 
@@ -23,9 +23,9 @@ internal class CalculatorMemoryTest {
     @Test
     internal fun `저장된 계산 수식을 가져올 수 있어야 한다`() {
         //given
-        val expression1 = CalculatorMemoryImpl.SAVE_FORMAT.format(Expression(listOf(3, Operator.Plus, 5)), 8)
-        val expression2 = CalculatorMemoryImpl.SAVE_FORMAT.format(Expression(listOf(4, Operator.Plus, 2)), 6)
-        calculatorMemory = CalculatorMemoryImpl(arrayListOf(expression1, expression2))
+        val expression1 = CalculatorMemory.SAVE_FORMAT.format(Expression(listOf(3, Operator.Plus, 5)), 8)
+        val expression2 = CalculatorMemory.SAVE_FORMAT.format(Expression(listOf(4, Operator.Plus, 2)), 6)
+        calculatorMemory = CalculatorMemory(arrayListOf(expression1, expression2))
         val expected = listOf(expression1, expression2)
 
         //when
